store user token in database
